Add DBus method to retrieve recv intf
Added code that will allow us to pass a transmitting interface name and
get back the receiving interface name and group number.
This is useful for the rfc interfaces that we generate, the perl
interface object that is created for the rfc interface needs to know
the real interface it is attached to.
Previously this was done by writing the mapping to a file and reading
it when the interface became active.
This change will let us call an RPC and get the name back. The RPC will
be written in the new python VCI component

Part of VRVDR-43653