Fix IPv6 Source address being removed on reload
IPv6 addresses in non-rfc mode end up getting added to the interface_t
as if they were the source address. This looks to be an ordering thing
where if the link-local vip netlink message appears before the interface
link-local generated address then it is added to the interface_t as the
source address, and then deleted later in the start up sequence such
that the interface_t is left without an address and keepalived thinks it
is down. This only happens on a config reload. In the first start up
everything works fine.
To work around this on a reload we delete the vip address before we
clear everything up, this forces the actual link-local to be added first
on the restart.

Closes:VRVDR-47349