Enhance IF_ISUP check to include lower layers
It is possible for an interface to be reported up even if the lower
physical or bond members are down. To get around this a start-delay
command was introduced, but this only works on start up of keepalived.

With further investigation there are some netlink flags and information
that can be used to determine the state of the lower levels. The
start-delay command may now be obsolete.

Introduced further netlink flag checks and added a member to the
interface_t structure to track the state of the IFLA_CARRIER property

Closes:VRVDR-49440