Bugfix/Stop syncgroup state 4 lockup
It is possible for sync groups to become locked in a transient
backup/state 4 state and never continue the FSM transitions. This is
caused by the wantstate of the individual VRRP groups being different.
This state is reproducible when VRRP is initially configured with a
single group in a sync group, and then a second group is added to the
sync-group (with the legacy configuration this second group is on the
same interface) in another commit. As a sync group with only one VRRP
group in it doesn't make sense the sync-group configuration for that
VRRP group is removed. On the second commit the sync group becomes
active but as the existing VRRP group was already in the MASTER state
and the new group starts it's FSM from scratch the total sync group gets
locked up.
To work around this problem on a reload we check that all the want
states of the individual VRRP groups are the same. If they are not the
same then we should go through all of the VRRP groups in the sync group
and set the want state of each to the lowest state seen in the sync
group.

Closes:VRVDR-47086