Bugfix/IPv6 Global IPAO restricted to RFC
Attempting to allow Global IPv6 addresses to be used for IP address
owner mode caused issues in non RFC configuration. When adding
additional VRRP configuration the reloading of the config file starts
the VRRP FSM which causes the checking of addresses on an interface. As
the VIP is on the interface this moves the group to IPAO mode
erroneously.
Added code that will check if we are in RFC mode and only then search
for Global IP addresses on the transmit interface to go into IPAO mode.

Closes:VRVDR-50153