Bugfix/Fix VRRPv2 SNMP timeout
When issuing an snmpwalk on vrrpOperTable the walk will only get so far
before getting locked up and timing out. The cause of this is the
var_len being incorrect for the variable returned.
Fixed two issues where this happened. One in vrrpOperTable and one in
vrrpAssoIpAddrTable (don't know why this one stopped a walk on
vrrpOperTable working).
Any upstream push of this should remove the IPv6 changes. I don't think
upstream uses IPv6 with VRRPv2

Closes:VRVDR-50365