Bugfix/Ensure router_id contains valid data
When keepalived reloads it's configuration there is a chance that
global_data->router_id will be set to NULL which causes a crash in the
SNMP code when trying to add this value to an SNMP trap.
Router_id will be NULL if a call to uname returns < 0 or if there is an
error calling getaddrinfo.
To mitigate this problem if the uname call is successful but the
getaddrinfo call isn't the uname.nodename member will be returned
If the uname call is unsuccessful there's probably bigger problems so
we'll log warnings and return the NULL value but we'll still attempt to
continue.
To do this we'll set the router_id to be the const string
"ROUTER NAME NOT FOUND" this prevents keepalived crashing.

Closes:VRVDR-50785