Bugfix/Reset prefix when deleting IP addresses
When a VRRP group moves to the MASTER state all the IPv4 addresses that
are to be added to the interface are checked against existing IP
addresses and their prefixes. If any of these VIPs that are to be added
overlap with an existing IP/prefix combination then the VIP's prefix is
set to /32. This is needed because if the IPs overlap then when one is
deleted the routing information is removed, this is a kernel thing.
Currently once the VIPs prefix has been set to /32 it can not be reset
to it's configured prefix.
This commit changes this behaviour such that when IP addresses are
deleted from an interface the keepalived structure for the IP address
resets the prefix length back to it's original length.
BACKUP instances will now always show the configured prefix.

Closes:VRVDR-46464