Bugfix/Sync carrier_is_up from netlink messages
carrier_is_up was getting set once at startup and wasn't being updated
on new netlink messages for when the interface went up of down.
Added code to update the ifp structure's carrier_is_up member at the
same time the ifp flags are updated when we receive netlink messages for
the interface.

Closes:VRVDR-50956