Bugfix/set state to BACKUP when cfg prio changes
Keepalived has a fast sync mechanism when it's configuration file is
reloaded to restart VRRP groups in the same state they were in before
the configuration was reloaded.
When lowering the priority of the current MASTER group below the
priority of a BACKUP group the newly changed group will come up as
MASTER first sending packets and sending an SNMP trap saying this router
is MASTER before receiving an advertisement from the new (correct)
MASTER and dropping into BACKUP state.
The sending of an SNMP trap interferes with tests so to avoid this side
effect this change set modifies the restarted state to be BACKUP if the
base priority in the old config and new config do not match.

Closes:VRVDR-51342