Add master down timer log reason
For most reasons behind a transition to MASTER state it's obvious the
cause, usually through other log messages around the transition. The one
change that isn't explicitly called out is when the group changes due to
the master down timer expiring.
For both VRRPv2 and VRRPv3 the vrrp->stats->master_reason is updated
when the timer expires, but only when keepalived is compiled with SNMP
RFCv3. DANOS/Vyatta is always compiled with this options.
Added code that checks vrrp->stats->master_reason to set a string with
the reason for the transition to MASTER state.

Closes:VRVDR-51191