Bugfix/Stop crash when intf not yet present
Keepalived expects all configured interfaces for groups to be present
when it starts up, if they don't it attempts to add the group to a
default interface (eth0 which doesn't exist on DANOS). This causes a
segfault when attempting to create dbus objects. The VRRP child process
then reloads and hits the issue again until the interface exists.
This only happens because the new VCI component starts Keepalived before
the VIF interfaces exist on the system.

Keepalived simply isn't built for this kind of event, this is at best a
work around of the issue and would need a complete rework of Keepalived
to get it to respond to an interface becoming active. Reworking in this
ancient version of the code is not worthwhile when it would need to
be done again for a more up to date version of Keepalived.

I tried various different methods of waiting for interfaces to become
present:
    * Introducing a sleep.
    	 keepalived is single threaded this just delays the crash.
    * Removing VRRP groups that don't have and active interface from the
      config.
      	With only one group configured removing it causes Keepalived to
	shutdown with a config error.
    * Being clever and trying to instantiate the VRRP instance with an
      empty interface_t instance and swapping it out when the interface
      exists.
        Too many areas of the code take for granted that the interface
	exists and it's structure is filled out completely causing
	errors in the log. Swapping out the structure didn't seem to
	work once the VRRP group was instantiated.
The code change in this patch is the only consistent method I've found
to work.

On config read if the interface we're looking for doesn't exist we
create the structure and set a new flag to indicate the interface
creation has been deferred.
For all VRRP groups that have a deferred interface we do not create
sockets for them or add them to the scheduler.
In the netlink code where we check for interfaces appearing if there is
already a structure for this interface we take a note of the deferred
status before we clear the structure and repopulate it.
If the previous deferred status was not present we schedule a reload
config.
The configuration is then reloaded and we continue on with the normal
setup and running of Keepalived.

This is only done for the one or two seconds before the VIF is created,
so I see no issue around things like last transition time/time taken to
reload and reread configuration.

Closes:VRVDR-51758