Allow multiple VRID on an intf when deferring intf
When using deferred interfaces all deferred interface indexes are the
same, this causes a config error when the same VRID is used on multiple
interfaces. A config error causes keepalived to exit and show commands
will report that VRRP is not running.

Added a check for reloads that will now only error if the old interface
structure or new interface structure haven't been deferred, the error
should still be reported if somehow two groups share the same VRID on a
configured interface.

Refreshed later patch

Closes:VRVDR-52186