Bugfix/Allow multiple groups on a deferred intf
The change to allow deferred interfaces has a bug where if the deferred
interface has multiple groups configured on it then the deferred
interface doesn't restart keepalived when it becomes active. This is due
to a hold over from the previous implementation.
Removed code that set the deferred status when a deferred interface
structure already exists.
Added a check for a VRRP structure being NULL which can potentially
cause a segfault.

Closes:VRVDR-52156