Bugfix/Multicast doesn't work when carrier down
When we attempt to create VMAC interfaces while the carrier protocol
is down all the setsockopt calls complete successfully, but the VRRP
multicast isn't added to the hardware filter.

Ideally we would only start the VRRP group's FSM once the interface
is completely up but keepalived doesn't currently have that
granularity to allow this. Instead we restart the VRRP process when the
carrier protocol is completely up.

Given the age of the vyatta keepalived version and the new scheduler in
later versions of keepalived it's not worth investing the time to make
the 1.2.24 scheduler react in this way. The work would have to be redone
for the new version of keepalived.

Closes:VRVDR-53740
(cherry picked from commit 60a4b15cf2cb0c9fc0934f751f0e25273ea10e77)