Bugfix/Fix unneeded vmac up on resignation
Using the advF variable in vrrp_restore_function as a method to resign
caused an unfortunate side effect in BACKUP state VRRP groups in a
sync-group. There is a branch of code that would flip RFC interfaces
into u/u when the advF variable was True (when resigning). Other code
branches bring the RFC interface down but when a group (G1) that is
BACKUP and in a sync-group that is moved to BACKUP - another VRRP
group (G2) in the sync-group quickly flips between MASTER and BACKUP -
G1 can follow a code path that doesn't bring the RFC interface down.

Changed the erroneous block of code to bring the RFC interface down to
only bring the interface down if preprocessor flags exist, the interface
exists, and the interface is up.

This modified patch 0193-Cleanup-RFC-Compliant-implementation.patch changing
the code block,
fix-rfc-mode-post-1.2.24.patch removed a whitespace change to the now defunct
code block,
terminate_forus_macvlan.patch has diffs moved around but no material changes.
Refreshed other patches for this code change.

Closes:VRVDR-54142