Bugfix/minimum changes to run acceptance tests
This is the minimum change set required to get keepalived v2.2.0 up and
running on a vyatta router.

* Updated debian/rules to use new tmp directory ability from upstream -
this has a counter part in vyatta-vrrp that creates and deletes this
temporary directory, removed debian.patchs/configure_tmp_file_path.patch

* Updated debian/patches/series to use the new names of some patches
(I've taken this uprev to standardise the patch names to use underscores
instead of hyphens and quilt diffs to include the name of the C
functions in the diff set. The C function names is accomplished by
adding QUILT_DIFF_OPTS="--show-c-function" to my quitrc)

* Removed debian/patches/security_fix_tmp_files.patch as this upstream
version contains that fix

* Removed debian/patches/defer_intf_creation.patch, upstream has a
different method of accomplishing this and a change in vyatta-vrrp takes
care of this

* Renamed fix_ipv6_always_using_v3 to vyatta_upstream_reversions.patch
this patch is a catch all for minor reversions from upstream, including:
 - allow IPv6 to be used with VRRPv3
 - print the count of VIPs in the data file
This patch wont be sent upstream

* Renamed vci_dbus_enhancements.patch to vyatta_dbus_rfc_additions.patch
this patch is used for a lot of VRRP show commands and to find the
interface that an RFC interface is configured on. Some minor changes to
some returned values until dynamic IPAO detection is turned on again.

* Renamed vyatta-if-additions.patch to vyatta_if_additions.patch and
modify it for upstream changes