Bugfix/Ignore own packets on restart
On occasion VRRP restarts fast enough to pick up a packet that the
instance previously sent. Usually this doesn't cause an issue but it can
be a problem if the new priority is lower than the previous priority
used and preemption is turned off. This packet can cause the new
instance to stay in BACKUP state and another VRRP router to take over
erroneously for the situation.

Changed the check for packet discarding to include packets that have a
priority greater or equal to our effective priority and that do not have
IPAO configured.

Closes:VRVDR-54397