Add extra data structures required by Vyatta
Vyatta requires some extra data for various features, instead of having
them buried in feature patches that later patches then rely on extract
them out into their own patch.
First of these data structures is the result of getifaddrs() that needs
to be updated every time an address is added or removed on the router.