Bugfix/Apply BPF to interface to receive VRRP adv
Background to this change:

Keepalived vanilla behaviour for created RFC interfaces is to send and
receive VRRP advertisements on this interface, or to send VRRP packets
from the base interface if use_xmit_base is configured. Sending and
receiving on an RFC interface is accomplished by having the interface
created and u/u (UP) on all routers in the VRRP group and installing the
Virtual Internet Protocol addresses (VIPs) on the Router that is acting
as the MASTER. BACKUP groups can receive the advertisements because the
interface is u/u and the interface has joined the VRRP Multicast Group,
this keeps the VRRP Finite State Machine (FSM) working as intended.

At the same time Vyatta expects the VRRP RFC interface to be A/D when
the router is in the BACKUP state of the VRRP FSM. When the interface is
A/D no advertisements are received so BACKUP routers transition to
MASTER then receive an advertisement and transition to BACKUP then
receive no advertisements and... This leads to a lot of state flipping.

Previous iterations of the Vyatta code used specific interface types and
would require a kernel change for upstream to take while also abusing
the use_xmit_base functionality (should really have been use_recv_base).
The specific interface type was removed thanks to the use of a Berkley
Packet Filter (BPF). This BPF is attached to the base interface on the
read socket and VRRP advertisements received there are passed to the
VRRP group, allowing the BACKUP routers to receive packets but to remain
A/D.

What this change does:
This commit updates the patch used to create the Vyatta specific
interface behaviour for the keepalived v2.2.0 release:
 * Dropped the need for use_xmit_base as it was confusing and we can get
   everything we need from the configured_ifp using the
   IF_BASE_IFP/IF_BASE_INDEX macros.
 * Added a function to bring a VMAC interface down via netlink.
 * Exposed this netlink down and the existing netlink up function in the
   vrrp_vmac header file.
 * Created the BPF and rewrote open_vrrp_read_socket() to use this instead
   of the upstream methodology.
 * Removed flags to recvmsg() the given ones cause an invalid argument
   error
 * Added further cmsg options when reading a packet from a socket.
     The BPF picks up packets that come from us as well so we should
     ignore them and any packets with a vlan tag.
     The packets sent to a VIF interface have the vlan tag stripped.
 * Updated the patch header with this explanation.

Changed debian/rules to compile in debug code, this makes debug easier
turning on with flags at start up when needed. Will update the DBus
debug code to use the in built debugs.