Bugfix/Stop incorrect removal of IP route with RFC
Merge stop_incorrect_ip_route_removal.patch into the
vyatta_rfc_changes.patch. This code changes the prefix used to add a VIP
if the configured prefix overlaps with an existing IP address, the new
prefix /32 stops the removal of IP route if the VIP is modified.

Merged add_effective_and_configured_prefix.patch into this patch as it's
a result of the stop_incorrect_ip_route_removal.patch changes.

Changed when update_interface_flags is called, should only be called
if the physical interface's flags change not the VMAC interface.