Add dynamic IP address owner mode detection
By the VRRP RFC any VRRP group that uses an IP address that is present on the
router should be placed into a special mode called IP address owner mode. This
mode changes the priority of the group to 255, ignoring the configured priority.
Currently Keepalived only uses this mode if it is explicitly configured in the
keepalived.conf file and the VIPs must be manually checked to be an address
owner.

For vyatta the keepalived.conf file is generated by python code and in the CLI
it is not possible to set the priority to 255, because of this we need to add
code that will dynamically check for IPAO mode.

Checking for IPAO mode should be done at initial configure and when new IP
addresses are added because of this.

To accomplish this I've added two new variables to the VRRP structure:
 * addr_owner - boolean for if the group is in IPAO mode
 * configured_priority - backup of configured priority, to be reinserted
 	if the group falls out of address owner mode
On initial configuration and when any IP address is added we check each VIP in
the VRRP group against the source address for the interface, if any of them
match then we set the priority to 255, backup the configured priority, and set
the address owner boolean to true.

There's an additional modification for IPv6. As well as checking the IPv6
Link Local source address we also check the IPv6 Global Addresses on the
interface for matching addresses.