Bugfix/Replace IPv6 cmsg code with IPv6 hdr checks
The cmsg code for IPv6 is unneeded as the information obtained from them
is found in the header which is passed to the receive call when using
AF_PACKET sockets and BPFs.
Replaced the cmsg code with code that parses the IPv6 header.
Updated the expected IPv6 packet length to include the IPv6 header
length (40)
Added ETH_P_ALL as the bind protocol, without this IPv6 sockets don't
bind after a config reload.
Removed the setsockopt bind as the existing function call does this.