Allow VRRP group BACKUP reset with DBus call
Currently there's no easy way to reset a VRRP group's state to BACKUP
from MASTER, this can be useful for forcing a failover for testing.
This patch introduces a DBus method call that can be used to reset
group state to BACKUP.
There's a side effect when a SyncGroup is used if one group changes
state all of the VRRP groups in the SyncGroup should switch state as
well. This requires an additional boolean argument to vrrp_sync_backup()
for other members of the SyncGroup to send a resign (zero priority)
packet and move to BACKUP.