Add IPv4 routes as trackable objects
Added IPv4 routes as a trackable object, tracking routes can be useful for
switching MASTER state if one group can reach a particular network.
To accomplish this we listen for netlink messages for routes being added or
deleted and adjust states and priorities as appropriate.

This addition follows a similar pattern as the existing upstream object tracking
with a structure containing the object we're tracking and the weighting for this
object, if any, being attached to the VRRP group that is tracking the object.
There's a deviation from the upstream code in how the complete set of objects
are tracked, instead of a list of all objects (Upstream) the objects are kept
in a Hashmap (provided by glib) and accessed using the route/mask combination
of the route. Apart from this change the flow is the same:

 * Seed initial value (from cold start or restart)
 * Listen for changes to object
   - Modify the is_up boolean
   - If the weight is nonzero modify priority and normalize
   - Else modify state to FAULT
 * Listen for changes and repeat

There are two vyatta specific changes for tracking routes.
The first is a modification to the existing netlink route code, Upstream is only
looking for routes that have been added be keepalived this existing code had
been hidden in an if branch and the new netlink route code added as the else
branch.
The second is zeroing the weight of tracked interfaces and tracked routes for
VRRP groups that are part of a Sync Group. This is to keep expected
functionality with previous releases, Sync Groups should only change states
not modify priority.