Add path-monitor instances as trackable objects
This is a vyatta specific patch as the path-monitor feature is closed source.

This patch provides functionality to track the state of a path-monitor and have
it affect the state or priority of a VRRP group. This is useful for when one
path to a destination is quicker than another. The details of path-monitors are
documented elsewhere, this change set is only concerned with the compliance
state of the configured path-monitor.

Path-monitors periodically advertise their compliance via DBus signals,
listening for these signals allows us to modify the VRRP group state.
A Path-monitor can be DELETED, COMPLIANT, or NON_COMPLIANT, due to the YANG
definitions for VRRP we can't track a Path-monitor that isn't configured and
a Path-monitor can not be deleted while it is being tracked. This simplifies
what we must listen for as any Path-monitor with a DELETED state can be ignored.

This change set follows the same design pattern as upstream tracked objects.
The tracked object is it's own structure with it's tracked state and a list of
VRRP groups that track this object.
Each VRRP group then has a list of structures that points to the tracked object
structure and contains the weighting to modify the VRRP Priority by if the
tracked object is UP or DOWN.
This seems a bit complex but makes it easier in practice to modify VRRP
priorities.

A Path-monitor is uniquely defined by a combination of a Monitor name and a
Policy name, the state can be queried directly using this combination. On
signal receipt the Monitor and Policy name can be compared to the values of
stored tracked object structures and then the compliance value in the signal
stored as well.

The state of the Path-monitor is fetched synchronously when the VRRP group is
first configured, then in the DBus code we set a callback for the Path-monitor
signal. As the state of the Path-monitor is sent periodically we only need to
take notice when the state changes, if we don't we end up repeatedly adding the
weighting to the VRRP priority.

The callback needs to run in the VRRP thread otherwise when we access the VRRP
data structures the code segfaults.

PartOf:VRVDR-33885