Refactor/Add carrier is up interface check patch
Checking for IFF_UP and IFF_RUNNING isn't enough to determine if
packets can be sent, some interface types also have a carrier protocol
that must be up before packets can be sent or received.

Relying on IFF_UP and IFF_RUNNING can cause a dual master on these
interfaces, IFF_LOWER_UP and checking IFLA_CARRIER in the RTM_NEWLINK
message allows us to wait for these interfaces to come up fully. This
can stop the need for delaying start up checking these variables means
we know when an interface is up and can send/receive.

Updated how interface flags are updated so the calculation on if a
change in state has happened includes the IFLA_CARRIER status. This
is helpful for VLAN interfaces where if the lower interface is down the
VRRP group stays up.