Bugfix/IPv6 RFC LL removed on master->backup chge
When an IPv6 RFC VRRP group moves out of the primary role the Link-local
IPv6 address from the lower interface is removed. This address is used
as the source address for VRRP packets, without this address the sendmsg
calls in vrrp_send_pkt fails with the ambiguous error invalid argument.

Stopping the removal doesn't seem to help with the error as binding to
the socket on a restart then fails. Instead when we move to master state
we read add the IP address before we add the VIP, if we add the VIPs
before hand then the LL VIP is taken as the source address which is also
wrong.

PartOf:VRVDR-33885