Refactor/Ignore base intf VRF in vmac creation
If the RFC VMAC interface is created outwith the default VRF then the
current Vyatta VRRP regression runs fail.
Added code that will allow the interface creation to be VRF aware based
on a config option. This will default to false, VMAC created in the
default VRF, to allow the existing regressions to complete. In the
future a YANG change would allow this to be configurable.

PartOf:VRVDR-33885