Bugfix/rfc preempt delay send zero pri
Previously we'd shutdown the VMAC RFC interface and then attempt to send
the zero priority advert, this patch moves the closing of the interface
to after the advert is sent.

Because of the way quilt works and the fact that there's no "new" code
just moving of existing code this commit appears to be just a code
refresh but there is a functional effect.

PartOf:VRVDR-33885