Bugfix/Use base interface to create DBus paths
When an interface was created at the same time as VRRP config with RFC
configuration was applied the deferred interface would mean the DBus
path for the VRRP group would use the RFC interface name.
This caused an issue with the show vrrp output which would display
default values but a show vrrp detail would show the correct
information.
Simplified how keepalived chooses the interface to use to create the
DBus path

Closes:VRVDR-56133