Bugfix/IPv6 BACKUP switch when VIP removed
When preempt was disabled on an IPv6 group and config was changed the
IPv6 group would incorrectly fall down into BACKUP. This was caused by
netlink messages from before the config reload being dealt with after
the reload. There is a check in the netlink code to set the VRRP group
to BACKUP if a VIP is removed, but this check didn't also check for the
preempt status.

I can't find a way to check the sequencing of the netlink messages to
discard those that are sent before a reload, so while it would be better
to ignore those messages this fix only drops a group to BACKUP if
preempt is enabled. This solves the symptoms of the issue.

Closes:VRVDR-56215