Bugfix/correct fault count for intf creation
Keepalived double counts an interface that has yet to be created as a
down fault, once at start up and then once again when the interface is
created, without a corresponding decrement for both events.
Added a new member to the interface_t structure that describes the
creation state of the interface (true or false), this is then used in
the interface tracking priority function. When the interface hasn't been
created yet the initialisation is skipped to avoid the double counting
described above.

Closes:VRVDR-56258