Bugfix/RFC intf deleted incorrectly with vrf-aware
If multiple VRRP groups were configured with RFCs and VRF aware on
configuration reload for VRF moving an RFC interface could be deleted
and recreated and then deleted again. This caused issue with show
commands and protocol operation.
The root cause was not checking if the old VRRP instance in a for loop
matched the new VRRP instance.
Instance vrrp-dp1-1 would have it's RFC interface deleted and recreated
but then when vrrp-dp2-2 was analysed the RFC interface for vrrp-dp1-1
would be deleted and then the code would try to recreate an RFC
interface for vrrp-dp2-2 that hadn't been deleted.
Fixed this by adding a string comparison to check if the old_vrrp
instance name matched the new vrrp instance name before checking for VRF
awareness changes between config.

Closes:VRVDR-56528